poster_service/test_replicate_service.py
"""
Test script for Replicate Florida Map Generation Service

This script tests the service structure and provides usage examples.
Run this to verify everything is set up correctly before using the API.
"""

import dotenv
import os
import sys
from pathlib import Path

dotenv.load_dotenv()

# Add services directory to path
sys.path.append(str(Path(__file__).parent / 'services'))

try:
    from services.replicate import FloridaMapGenerator, MapStyle
    print("✅ Successfully imported Replicate service")
except ImportError as e:
    print(f"❌ Failed to import Replicate service: {e}")
    sys.exit(1)


def test_service_structure():
    """Test that the service structure is correct"""
    print("\n🔍 Testing service structure...")
    
    # Test enum values
    print(f"📊 Available AI Map Styles: {[style.value for style in MapStyle]}")
    
    # Test service initialization (without API token)
    try:
        # This should fail without API token - that's expected
        generator = FloridaMapGenerator()
        print("❌ Unexpected: Service initialized without API token")
    except ValueError as e:
        print(f"✅ Expected error without API token: API token required")
    
    print("✅ Service structure test passed")


def test_prompt_generation():
    """Test that prompts are properly structured"""
    print("\n🎨 Testing prompt generation...")
    
    # Test that we can access style prompts
    from services.replicate import FloridaMapGenerator
    
    for style in MapStyle:
        prompt_config = FloridaMapGenerator.STYLE_PROMPTS[style]
        print(f"\n📝 {style.value.upper()} Style:")
        print(f"  Keywords: {prompt_config['style_keywords']}")
        print(f"  Prompt length: {len(prompt_config['prompt'])} characters")
        
        # Check that coastal spacing is mentioned
        if "coastal" in prompt_config['prompt'].lower() or "margin" in prompt_config['prompt'].lower():
            print("  ✅ Includes coastal spacing optimization")
        else:
            print("  ⚠️  May need more coastal spacing optimization")
    
    print("\n✅ Prompt generation test passed")


def show_usage_examples():
    """Show usage examples for the service"""
    print("\n📖 Usage Examples:")
    print("=" * 50)
    
    print("""
1. SETUP - Get your Replicate API token:
   • Go to: https://replicate.com/account/api-tokens
   • Create a new token
   • Set environment variable:
     export REPLICATE_API_TOKEN=your_token_here

2. BASIC USAGE:
   from services.replicate import FloridaMapGenerator, MapStyle
   
   generator = FloridaMapGenerator()
   
   # Generate a single map
   map_path = generator.generate_map_template(
       style=MapStyle.CLASSIC,
       width=1024,
       height=1024,
       save_path="florida_classic.png"
   )

3. GENERATE ALL STYLES:
   all_maps = generator.generate_all_styles(
       output_dir="templates",
       width=1024,
       height=1024
   )

4. CUSTOM PROMPT:
   custom_map = generator.generate_custom_map(
       custom_prompt="Neon cyberpunk Florida map, glowing edges",
       style_name="cyberpunk",
       save_path="florida_cyberpunk.png"
   )

5. FULL AI PIPELINE:
   from ai_poster_pipeline import AIPosterPipeline
   
   pipeline = AIPosterPipeline()
   poster = pipeline.generate_single_poster(
       ai_style=MapStyle.VINTAGE,
       poster_title="Florida Surf Breaks - Vintage Collection"
   )
""")


def show_optimization_features():
    """Show how the service is optimized for poster creation"""
    print("\n🎯 Optimization Features for Surf Break Posters:")
    print("=" * 50)
    
    print("""
✅ COASTAL SPACING OPTIMIZATION:
   • All prompts include 15% minimum margins around coastline
   • Text label placement areas are specifically requested
   • White space/background space is maximized for readability

✅ STYLE-SPECIFIC OPTIMIZATIONS:
   • Classic: Clean cartographic style with minimal distractions
   • Vintage: Aged paper with wide margins for retro text
   • Minimalist: Maximum negative space for clean labeling
   • Watercolor: Soft edges that don't compete with text
   • Art Deco: Geometric elements that complement text placement

✅ TECHNICAL OPTIMIZATIONS:
   • High resolution (1024x1024) for crisp text rendering
   • Professional color palettes that ensure text contrast
   • No pre-existing text/labels that would conflict
   • Centered state positioning for balanced layout

✅ INTEGRATION READY:
   • Direct compatibility with existing poster service
   • Automatic style mapping between AI and poster styles
   • Seamless pipeline from AI generation to final poster
""")


def main():
    """Run all tests"""
    print("🧪 Replicate Service Test Suite")
    print("=" * 40)
    
    # Test service structure
    test_service_structure()
    
    # Test prompt generation
    test_prompt_generation()
    
    # Show usage examples
    show_usage_examples()
    
    # Show optimization features
    show_optimization_features()
    
    # Final status
    print("\n" + "=" * 50)
    if os.getenv('REPLICATE_API_TOKEN'):
        print("🎉 Ready to generate AI maps! Your API token is set.")
        print("📝 Run 'python services/replicate.py' to test map generation")
        print("🚀 Run 'python ai_poster_pipeline.py' for the full pipeline")
    else:
        print("⚠️  Set REPLICATE_API_TOKEN to start generating AI maps")
        print("🔗 Get your token: https://replicate.com/account/api-tokens")
    
    print("✅ All tests passed! Service is ready for use.")


if __name__ == "__main__":
    main() 