poster_service/requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# AI and Image Processing
replicate>=0.25.0
pillow>=10.0.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
